requirements.txt
blake3
specific_import
psutil
toposort
flake8
autopep8
rope
rich-argparse
cchardet
openpyxl
sqlalchemy==1.4.32
black
command_runner==1.4.0
pathvalidate
jpype1==1.4.0
JayDeBeApi==1.2.3
sqlite-utils
PyYAML==6.0
typing-extensions
pipgrip
frictionless[sql] --pre
cryptography>=36.0.0
install-jdk
maven-artifact
dulwich